refactored markdown parsing functions into a separate file
